about a billion small to medium fixes

+++ helpers/log.js
@@ -5,9 +5,14 @@ const fs = require('fs');
 module.exports = async (error, client, rawError) => {
     console.log(`[ERROR] ${error}`, rawError)
 
-    fs.appendFileSync('errorLog.txt', `${rawError ? rawError.stack : ""}`);
-    const attachment = new AttachmentBuilder('errorLog.txt');
     const user = await client.users.fetch(ownerId);
-    await user.send({ content: `[ERROR] ${error}${rawError ? "" : "no stack trace!"}`, files: [attachment] })
-    fs.rmSync('errorLog.txt');
+
+    if (rawError && rawError.stack) {
+        fs.appendFileSync('errorLog.txt', `${rawError ? rawError.stack : ""}`);
+        const attachment = new AttachmentBuilder('errorLog.txt');
+        await user.send({ content: `[ERROR] ${error}`, files: [attachment] })
+        fs.rmSync('errorLog.txt');
+    } else {
+        await user.send(`[ERROR] ${error}`);
+    }
 }
